feat(phase0): MER-35 commit bpf2go bindings and close Phase-0 sign-off
Generate and commit vmlinux.h plus counter/tcingress/tcegress bpf2go artifacts,
run the four 5.15 exit gates in Lima (verify-gen, prog_test_run, ring decode,
PERCPU readback), and record Phase 1 entry APPROVED in PHASE0_REVIEW with
docs/PHASE0_GATE_EVIDENCE.log. Adds macOS Lima delegation scripts and fixes
ringbuf BTF extraction plus tc compile warnings for the codegen path.

+++ test/bpf/verdict_test.go
@@ -203,7 +203,7 @@ func TestVerdictMatrixMatchesReferenceEvaluator(t *testing.T) {
 			if !ok {
 				t.Fatalf("unsupported expected action in current matrix: %d", want.Action)
 			}
-			if got != wantTCAct {
+			if got != uint32(wantTCAct) {
 				t.Fatalf("verdict mismatch: kernel=%d expected(tc)=%d expected(action)=%d", got, wantTCAct, want.Action)
 			}
 		})
